Add 'clock and 'journal modes to the journal function
In the 'clock mode, you can either jump to the currently clocking
item, if there is one, or be prompted to choose from one of the recent
clocking items.

In the 'journal mode, simply show a capture buffer with the
template. One can call finished with a universal argument to jump to
the captured entry on finish, or call the pc/journal function with no
mode argument to jump to "today" in the journal.